+++ deepinv/tests/test_generators.py
@@ -34,7 +34,6 @@ MIXTURES = list(itertools.combinations(GENERATORS, 2))
 SIZES = [(5, 5), (6, 6)]
 NUM_CHANNELS = [1, 3]
 
-
 # MRI Generators
 C, T, H, W = 2, 12, 256, 512
 MRI_GENERATORS = ["gaussian", "random", "uniform", "poly"]
@@ -50,15 +49,21 @@ INPAINTING_IMG_SIZES = [
 ]  # (C,H,W), (C,M), (C,T,H,W)
 INPAINTING_GENERATORS = ["bernoulli", "gaussian", "multiplicative"]
 
-# All devices to test
-DEVICES = ["cpu"]
-if torch.cuda.is_available():
-    DEVICES.append("cuda")
-
 DTYPES = [torch.float32, torch.float64]
 
 
-def find_generator(name, size, num_channels, device, dtype, psf_size=None):
+# Fixture returns either None or a torch.Generator on the specified device, to test both cases:
+# 1. an existing generator is passed to the physics generator
+# 2. the physics generator creates its own default generator (according to the device values)
+# with this fixture, generator.rng.device and device always match.
+@pytest.fixture(params=[None, pytest.param("device", marks=pytest.mark.indirect)])
+def rng(request, device):
+    if request.param == "device":
+        return torch.Generator(device=device).manual_seed(0)
+    return None
+
+
+def find_generator(name, size, num_channels, device, dtype, psf_size=None, rng=None):
     r"""
     Chooses operator
 
@@ -73,38 +78,65 @@ def find_generator(name, size, num_channels, device, dtype, psf_size=None):
         keys = ["filter"]
     elif name == "MotionBlurGenerator":
         g = dinv.physics.generator.MotionBlurGenerator(
-            psf_size=size, num_channels=num_channels, device=device, dtype=dtype
+            psf_size=size,
+            num_channels=num_channels,
+            device=device,
+            dtype=dtype,
+            rng=rng,
         )
         keys = ["filter"]
     elif name == "DiffractionBlurGenerator":
         g = dinv.physics.generator.DiffractionBlurGenerator(
-            psf_size=size, device=device, num_channels=num_channels, dtype=dtype
+            psf_size=size,
+            device=device,
+            num_channels=num_channels,
+            dtype=dtype,
+            rng=rng,
         )
         keys = ["filter", "coeff", "pupil"]
     elif name == "ProductConvolutionBlurGenerator":
         g = dinv.physics.generator.ProductConvolutionBlurGenerator(
             psf_generator=dinv.physics.generator.DiffractionBlurGenerator(
-                psf_size=size, device=device, num_channels=num_channels, dtype=dtype
+                psf_size=size,
+                device=device,
+                num_channels=num_channels,
+                dtype=dtype,
+                rng=rng,
             ),
             img_size=512,
             n_eigen_psf=10,
             device=device,
             dtype=dtype,
+            rng=rng,
         )
         keys = ["filters", "multipliers"]
     elif name == "DownsamplingGenerator":
         g = dinv.physics.generator.DownsamplingGenerator(
-            filters=["bilinear", "bicubic", "gaussian"], factors=[2, 4]
+            filters=["bilinear", "bicubic", "gaussian"],
+            factors=[2, 4],
+            rng=rng,
+            device=device,
+            dtype=dtype,
         )
         keys = ["filters", "factors"]
     elif name == "DownsamplingGenerator2":
         g = dinv.physics.generator.DownsamplingGenerator(
-            filters=["bilinear", "bicubic", "gaussian"], factors=[2], psf_size=psf_size
+            filters=["bilinear", "bicubic", "gaussian"],
+            factors=[2],
+            psf_size=psf_size,
+            rng=rng,
+            device=device,
+            dtype=dtype,
         )
         keys = ["filters", "factors"]
     elif name == "DownsamplingGenerator4":
         g = dinv.physics.generator.DownsamplingGenerator(
-            filters=["bilinear", "bicubic", "gaussian"], factors=[4], psf_size=psf_size
+            filters=["bilinear", "bicubic", "gaussian"],
+            factors=[4],
+            psf_size=psf_size,
+            rng=rng,
+            device=device,
+            dtype=dtype,
         )
         keys = ["filters", "factors"]
     elif name == "DownsamplingGenerator[2, 4]":
@@ -112,13 +144,16 @@ def find_generator(name, size, num_channels, device, dtype, psf_size=None):
             filters=["bilinear", "bicubic", "gaussian"],
             factors=[2, 4],
             psf_size=psf_size,
+            rng=rng,
+            device=device,
+            dtype=dtype,
         )
         keys = ["filters", "factors"]
     elif name == "SigmaGenerator":
-        g = dinv.physics.generator.SigmaGenerator(device=device, dtype=dtype)
+        g = dinv.physics.generator.SigmaGenerator(device=device, dtype=dtype, rng=rng)
         keys = ["sigma"]
     elif name == "GainGenerator":
-        g = dinv.physics.generator.GainGenerator(device=device, dtype=dtype)
+        g = dinv.physics.generator.GainGenerator(device=device, dtype=dtype, rng=rng)
         keys = ["gain"]
     else:
         raise Exception("The generator chosen doesn't exist")
@@ -128,14 +163,15 @@ def find_generator(name, size, num_channels, device, dtype, psf_size=None):
 @pytest.mark.parametrize("name", GENERATORS)
 @pytest.mark.parametrize("size", SIZES)
 @pytest.mark.parametrize("num_channels", NUM_CHANNELS)
-@pytest.mark.parametrize("device", DEVICES)
 @pytest.mark.parametrize("dtype", DTYPES)
-def test_shape(name, size, num_channels, device, dtype):
+def test_shape(name, size, num_channels, device, dtype, rng):
     r"""
     Tests generators shape.
     """
 
-    generator, size, keys = find_generator(name, size, num_channels, device, dtype)
+    generator, size, keys = find_generator(
+        name, size, num_channels, device, dtype, rng=rng
+    )
     batch_size = 4
 
     params = generator.step(batch_size=batch_size)
@@ -147,14 +183,13 @@ def test_shape(name, size, num_channels, device, dtype):
 
 
 @pytest.mark.parametrize("name", GENERATORS)
-@pytest.mark.parametrize("device", DEVICES)
 @pytest.mark.parametrize("dtype", DTYPES)
-def test_generation_newparams(name, device, dtype):
+def test_generation_newparams(name, device, dtype, rng):
     r"""
     Tests generators' ability to generate new parameters at each step.
     """
     size = (32, 32)
-    generator, size, _ = find_generator(name, size, 1, device, dtype)
+    generator, size, _ = find_generator(name, size, 1, device, dtype, rng=rng)
     batch_size = 1
 
     if name == "GaussianBlurGenerator":
@@ -176,14 +211,13 @@ def test_generation_newparams(name, device, dtype):
 
 
 @pytest.mark.parametrize("name", GENERATORS)
-@pytest.mark.parametrize("device", DEVICES)
 @pytest.mark.parametrize("dtype", DTYPES)
-def test_generation_seed(name, device, dtype):
+def test_generation_seed(name, device, dtype, rng):
     r"""
     Tests generators consistency with the same random seed.
     """
     size = (32, 32)
-    generator, size, _ = find_generator(name, size, 1, device, dtype)
+    generator, size, _ = find_generator(name, size, 1, device, dtype, rng=rng)
     batch_size = 1
 
     if name == "GaussianBlurGenerator":
@@ -207,14 +241,13 @@ def test_generation_seed(name, device, dtype):
 @pytest.mark.parametrize(
     "name", sorted(set(GENERATORS).difference(set(["ProductConvolutionBlurGenerator"])))
 )
-@pytest.mark.parametrize("device", DEVICES)
 @pytest.mark.parametrize("dtype", [torch.float64])
-def test_average(name, device, dtype):
+def test_average(name, device, dtype, rng):
     r"""
     Tests generators average.
     """
     size = (5, 5)
-    generator, size, _ = find_generator(name, size, 1, device, dtype)
+    generator, size, _ = find_generator(name, size, 1, device, dtype, rng=rng)
     # Set generator seed for reproducibility
     generator.rng_manual_seed(0)
 
@@ -237,11 +270,10 @@ def test_average(name, device, dtype):
 
 
 @pytest.mark.parametrize("num_channels", NUM_CHANNELS)
-@pytest.mark.parametrize("device", DEVICES)
 @pytest.mark.parametrize("dtype", DTYPES)
 @pytest.mark.parametrize("psf_size", [None, (31, 31)])
 @pytest.mark.parametrize("fact", [None, 2, 4, [2, 4]])
-def test_downsampling_generator(num_channels, device, dtype, psf_size, fact):
+def test_downsampling_generator(num_channels, device, dtype, psf_size, fact, rng):
     r"""
     Test downsampling generator.
     This test is different from the above ones because we do not generate a random kernel at each iteration, but
@@ -265,6 +297,7 @@ def test_downsampling_generator(num_channels, device, dtype, psf_size, fact):
         device,
         dtype,
         psf_size=psf_size,
+        rng=rng,
     )
 
     batch_size = (
@@ -278,7 +311,11 @@ def test_downsampling_generator(num_channels, device, dtype, psf_size, fact):
     else:
         params = generator.step(batch_size=batch_size, seed=1)
 
-        x = torch.randn((batch_size, num_channels, size[0], size[1])).to(device)
+        x = torch.randn(
+            (batch_size, num_channels, size[0], size[1]),
+            generator=generator.rng,
+            device=device,
+        )
         y = physics(x, **params)
 
         assert y.shape[-1] == x.shape[-1] // params["factor"].unique().item()
@@ -297,22 +334,39 @@ def batch_size():
     return 2
 
 
-def choose_mri_generator(generator_name, img_size, acc, center_fraction):
+def choose_mri_generator(generator_name, img_size, acc, center_fraction, device, rng):
     if generator_name == "gaussian":
         g = GaussianMaskGenerator(
-            img_size, acceleration=acc, center_fraction=center_fraction
+            img_size,
+            acceleration=acc,
+            center_fraction=center_fraction,
+            rng=rng,
+            device=device,
         )
     elif generator_name == "random":
         g = RandomMaskGenerator(
-            img_size, acceleration=acc, center_fraction=center_fraction
+            img_size,
+            acceleration=acc,
+            center_fraction=center_fraction,
+            rng=rng,
+            device=device,
         )
     elif generator_name == "uniform":
         g = EquispacedMaskGenerator(
-            img_size, acceleration=acc, center_fraction=center_fraction
+            img_size,
+            acceleration=acc,
+            center_fraction=center_fraction,
+            rng=rng,
+            device=device,
         )
     elif generator_name == "poly":
         g = PolyOrderMaskGenerator(
-            img_size, acceleration=acc, center_fraction=center_fraction, poly_order=2
+            img_size,
+            acceleration=acc,
+            center_fraction=center_fraction,
+            poly_order=2,
+            rng=rng,
+            device=device,
         )
     return g
 
@@ -321,8 +375,12 @@ def choose_mri_generator(generator_name, img_size, acc, center_fraction):
 @pytest.mark.parametrize("img_size", MRI_IMG_SIZES)
 @pytest.mark.parametrize("acc", MRI_ACCELERATIONS)
 @pytest.mark.parametrize("center_fraction", MRI_CENTER_FRACTIONS)
-def test_mri_generator(generator_name, img_size, batch_size, acc, center_fraction):
-    generator = choose_mri_generator(generator_name, img_size, acc, center_fraction)
+def test_mri_generator(
+    generator_name, img_size, batch_size, acc, center_fraction, device, rng
+):
+    generator = choose_mri_generator(
+        generator_name, img_size, acc, center_fraction, device, rng
+    )
     # test across different accs and center fracations
     H, W = img_size[-2:]
     assert W // generator.acc == (generator.n_lines + generator.n_center)
@@ -400,7 +458,6 @@ def choose_inpainting_generator(name, img_size, split_ratio, pixelwise, device, 
 @pytest.mark.parametrize("img_size", INPAINTING_IMG_SIZES)
 @pytest.mark.parametrize("pixelwise", (False, True))
 @pytest.mark.parametrize("split_ratio", (0.5,))
-@pytest.mark.parametrize("device", DEVICES)
 def test_inpainting_generators(
     generator_name, batch_size, img_size, pixelwise, split_ratio, device, rng
 ):
@@ -469,7 +526,7 @@ def test_inpainting_generators(
     mask3 = gen.step(batch_size=batch_size, input_mask=input_mask, seed=0)["mask"]
     correct_ratio(
         mask3.sum() / input_mask.sum(),
-        atol=1e-2 if generator_name != "multiplicative" else 2e-1,
+        atol=1e-2 if generator_name != "multiplicative" else 2.5e-1,
     )
     correct_pixelwise(mask3)
 
@@ -485,17 +542,21 @@ def test_inpainting_generators(
 
 
 @pytest.mark.parametrize("num_channels", NUM_CHANNELS)
-@pytest.mark.parametrize("device", DEVICES)
 @pytest.mark.parametrize("dtype", DTYPES)
-@pytest.mark.parametrize("rng", [torch.Generator().manual_seed(0)])
 def test_inpainting_generator_random_ratio(num_channels, device, dtype, rng):
     # NOTE elements of this test are now redundant given above tests
     size = (100, 100)  # we take it large to have significant statistical numbers after
-    physics = dinv.physics.Inpainting((num_channels, size[0], size[1]), 0.9)
+    physics = dinv.physics.Inpainting(
+        (num_channels, size[0], size[1]), 0.9, device=device
+    )
 
     split_ratio = 0.6
     generator = dinv.physics.generator.BernoulliSplittingMaskGenerator(
-        (num_channels, size[0], size[1]), split_ratio=split_ratio
+        (num_channels, size[0], size[1]),
+        split_ratio=split_ratio,
+        device=device,
+        dtype=dtype,
+        rng=rng,
     )
     batch_size = 2
     params = generator.step(batch_size=batch_size)
@@ -507,7 +568,11 @@ def test_inpainting_generator_random_ratio(num_channels, device, dtype, rng):
     assert abs(experimental_split_ratio.item() - split_ratio) < 1e-2
 
     # check forward
-    x = torch.randn((batch_size, num_channels, size[0], size[1]), generator=rng)
+    x = torch.randn(
+        (batch_size, num_channels, size[0], size[1]),
+        generator=generator.rng,
+        device=device,
+    )
     y = physics(x, **params)
     experimental_split_ratio_obs = 1 - (y[0] == 0).sum() / y[0].numel()
     assert torch.allclose(
@@ -523,6 +588,8 @@ def test_inpainting_generator_random_ratio(num_channels, device, dtype, rng):
         random_split_ratio=True,
         min_split_ratio=min_split_ratio,
         max_split_ratio=max_split_ratio,
+        device=device,
+        rng=rng,
     )
     batch_size = 2
     params = generator.step(batch_size=batch_size, seed=0)
@@ -530,7 +597,11 @@ def test_inpainting_generator_random_ratio(num_channels, device, dtype, rng):
     mask = params["mask"]
     assert mask.shape == (batch_size, num_channels, size[0], size[1])
 
-    x = torch.randn((batch_size, num_channels, size[0], size[1]), generator=rng)
+    x = torch.randn(
+        (batch_size, num_channels, size[0], size[1]),
+        generator=generator.rng,
+        device=device,
+    )
     y = physics(x, **params)
 
     list_exp_split_ratio = []
@@ -574,7 +645,7 @@ def test_string_seed():
 @pytest.mark.parametrize("convention", ["noll", "ansi"])
 @pytest.mark.parametrize("is_3d", [True, False])
 def test_diffraction_generator(
-    device, apodize, random_rotate, num_channels, convention, is_3d
+    device, apodize, random_rotate, num_channels, convention, is_3d, rng
 ):
     r"""
     Test diffraction generator.
@@ -595,6 +666,7 @@ def test_diffraction_generator(
             random_rotate=random_rotate,
             dtype=dtype,
             pupil_size=pupil_size,
+            rng=rng,
         )
 
     else:
@@ -610,6 +682,7 @@ def test_diffraction_generator(
             random_rotate=random_rotate,
             dtype=dtype,
             pupil_size=pupil_size,
+            rng=rng,
         )
 
     batch_sizes = (1, 2)
@@ -649,7 +722,6 @@ def test_diffraction_generator(
         )
         for key in params.keys():
             assert not torch.allclose(params[key], params3[key])
-
 
 @pytest.mark.parametrize("dim", [1, 2, 3])
 @pytest.mark.parametrize("isotropic", [True, False])
@@ -865,3 +937,34 @@ def test_gaussian_blur_generator(device, dim, isotropic, batch_size, num_channel
         )
         f_angle_iso = p_angle_iso["filter"]
         assert torch.allclose(f_angle_iso[0], f_angle_iso[1])
+
+@pytest.mark.parametrize("generators", MIXTURES)
+@pytest.mark.parametrize("size", SIZES)
+@pytest.mark.parametrize("num_channels", NUM_CHANNELS)
+@pytest.mark.parametrize("dtype", DTYPES)
+@pytest.mark.parametrize("use_batch_sampling", [True, False])
+def test_generator_mixture(
+    generators, size, num_channels, dtype, use_batch_sampling, device, rng
+):
+
+    generator_pair = []
+    for name in generators:
+        g, _, _ = find_generator(name, size, num_channels, device, dtype, rng=rng)
+        generator_pair.append(g)
+
+    mixture = dinv.physics.generator.GeneratorMixture(
+        generator_pair,
+        [0.5, 0.5],
+        use_batch_sampling=use_batch_sampling,
+        device=device,
+        rng=rng,
+        verbose=True,
+    )
+
+    params = mixture.step(batch_size=4, seed=0)
+    assert isinstance(params, dict)
+
+    # Check the set keys of produced by the mixture are the same as the keys of the individual generators
+    assert set(params.keys()).intersection(
+        set.union(*[set(g.step(batch_size=1, seed=0).keys()) for g in generator_pair])
+    ) == set(params.keys())